requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-mptt>=0.14.0
django-ckeditor>=6.5.0
channels>=4.0.0
channels-redis>=4.0.0
redis>=4.5.0
asgiref>=3.6.0
daphne>=4.0.0
whitenoise>=6.4.0
gunicorn>=21.0.0
